Finalise uuid generation on process manifest task (#87)
Moves the execution order of the BugsnagManifestTask so that it 
completes before the processManifest completes. This means that 
changes made to the manifest such as build UUID generation are 
written to the file at the appropriate time. Otherwise the Manifest in 
the build directory and packaged APK may differ in their contents.