fix: ensure that build UUIDs are generated before the manifest is bundled
For App Bundles, the plugin could write to the manifest in bundle_manifest after it had been
packaged into a bundle. This meant the build UUID was not inserted and therefore crashes would not
deobfuscated properly, in projects where shrinkResources was enabled, which altered the task
ordering of the bundleResources task. This fix ensures that the bundleResources task has a
dependency on the manifest manipulation task so it will be run first.