Apply requestTimeoutMs as an idle timeout
This reinstates the previous (v4) behaviour of the timeout — it
used to apply as an idle timeout (i.e. fail if no activity has
happened in x ms) but 'callTimeout' is a hard time limit on a
request

This means that a request had 'requestTimeoutMs' ms to complete
in its entirety and, even if it was still uploading, we would kill
the request if this was exceeded

Now we use the read/write/connect timeouts, which are idle timeouts
and don't apply a callTimeout at all. This ensures that active
requests are allowed to continue, but we will still fail requests
that idle for too long