Make BuildUuid a settable extension property instead
This consolidates build uuid API usage in the manifest patching task and also exposes it as a property on the bugsnag plugin extension (with the current randomized behavior as the default). This allows consumers that do have their own deterministic heuristic to be able to set their own in a way that aligns with their own reproducibility/compliance requirements.